+++ doc/eslint.config_partial.mjs
@@ -1,4 +1,4 @@
-import { requireEslintTool } from '../tools/eslint.config_utils.mjs';
+import { requireEslintTool } from '../tools/eslint/eslint.config_utils.mjs';
 
 const globals = requireEslintTool('globals');
