+++ eslint.config.mjs
@@ -5,13 +5,14 @@ import benchmarkConfig from './benchmark/eslint.config_partial.mjs';
 import docConfig from './doc/eslint.config_partial.mjs';
 import libConfig from './lib/eslint.config_partial.mjs';
 import testConfig from './test/eslint.config_partial.mjs';
-import toolsConfig from './tools/eslint.config_partial.mjs';
+import toolsConfig from './tools/eslint/eslint.config_partial.mjs';
 import {
   noRestrictedSyntaxCommonAll,
   noRestrictedSyntaxCommonLib,
   requireEslintTool,
   resolveEslintTool,
-} from './tools/eslint.config_utils.mjs';
+} from './tools/eslint/eslint.config_utils.mjs';
+import nodeCore from './tools/eslint/eslint-plugin-node-core.js';
 
 const js = requireEslintTool('@eslint/js');
 const babelEslintParser = requireEslintTool('@babel/eslint-parser');
@@ -20,7 +21,6 @@ const jsdoc = requireEslintTool('eslint-plugin-jsdoc');
 const markdown = requireEslintTool('eslint-plugin-markdown');
 const stylisticJs = requireEslintTool('@stylistic/eslint-plugin-js');
 
-const nodeCore = requireEslintTool('eslint-plugin-node-core');
 nodeCore.RULES_DIR = fileURLToPath(new URL('./tools/eslint-rules', import.meta.url));
 
 // The Module._resolveFilename() monkeypatching is to make it so that ESLint is able to
