+++ lib/eslint.config_partial.mjs
@@ -3,7 +3,7 @@
 import {
   noRestrictedSyntaxCommonAll,
   noRestrictedSyntaxCommonLib,
-} from '../tools/eslint.config_utils.mjs';
+} from '../tools/eslint/eslint.config_utils.mjs';
 
 export default [
   {
