+++ tools/eslint/eslint-plugin-node-core.js
@@ -13,7 +13,7 @@ module.exports = {
     if (!cache) {
       cache = {};
       const files = fs.readdirSync(RULES_DIR)
-        .filter(filename => filename.endsWith('.js'))
+        .filter((filename) => filename.endsWith('.js'));
       for (const file of files) {
         const name = file.slice(0, -3);
         cache[name] = require(path.resolve(RULES_DIR, file));
