+++ tools/eslint/eslint.config_partial.mjs
